requirements.txt
Django>=5.0
python-chess